refactor: Moved annotation table to a separate component
